requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argon2==0.1.10
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.12
django-extensions==3.1.5
djangorestframework==3.12.4
djongo==1.3.7
drf-yasg==1.20.0
filelock==3.19.1
fsspec==2025.9.0
grpcio==1.76.0
hydra-core==1.3.2
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.0
inflection==0.5.1
iopath==0.1.10
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.9
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pymongo==3.11.4
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scipy==1.15.3
simsimd==6.5.3
sqlparse==0.2.4
stringzilla==4.2.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.10.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
git+https://github.com/USER/REPO_NAME.git@TAG_HOẶC_BRANCH_HOẶC_COMMIT#egg=PACKAGE_NAME